requirements.txt
APScheduler==3.11.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
mariadb==1.1.14
MarkupSafe==3.0.3
packaging==25.0
python-ulid==3.1.0
requests==2.32.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
Werkzeug==3.1.4
